requirements.txt
# 
# RLHF Preference Platform - Requirements
# 

# API & Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9        # file uploads

# UI
gradio>=4.20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0                # optional: migrations

# Data / Metrics
pandas>=2.0.0
numpy>=1.26.0

# ML / Reward Model
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.27.0

# Optional: advanced reward training with TRL
# trl>=0.7.0

# Config
python-dotenv>=1.0.0

# HTTP (Gradio -> FastAPI calls)
requests>=2.31.0

# Optional: OpenAI-compatible backend
# openai>=1.12.0

# Groq Cloud API (fast LLM inference)
groq>=0.9.0
